refactor(plugins): move HTTP, Serialization, and Databases to hexagonal layers\n\nCo-Authored-By: Claude <noreply@anthropic.com>
