+++ src/test/kotlin/io/sdkman/state/HealthCheckApiSpec.kt
@@ -20,7 +20,7 @@ import io.sdkman.state.adapter.secondary.persistence.PostgresVersionRepository
 import io.sdkman.state.application.service.TagServiceImpl
 import io.sdkman.state.application.service.VersionServiceImpl
 import io.sdkman.state.config.configureAppConfig
-import io.sdkman.state.plugins.configureDatabase
+import io.sdkman.state.config.configureDatabase
 import io.sdkman.state.support.withCleanDatabase
 import kotlinx.serialization.json.Json
